Rename no-noninteractive-element-handlers to no-noninteractive-element-interactions
