+++ src/index.js
@@ -22,7 +22,7 @@ module.exports = {
     'no-access-key': require('./rules/no-access-key'),
     'no-autofocus': require('./rules/no-autofocus'),
     'no-distracting-elements': require('./rules/no-distracting-elements'),
-    'no-noninteractive-element-handlers': require('./rules/no-noninteractive-element-handlers'),
+    'no-noninteractive-element-interactions': require('./rules/no-noninteractive-element-interactions'),
     'no-onchange': require('./rules/no-onchange'),
     'no-redundant-roles': require('./rules/no-redundant-roles'),
     'no-static-element-interactions': require('./rules/no-static-element-interactions'),
@@ -60,7 +60,7 @@ module.exports = {
         'jsx-a11y/no-access-key': 'error',
         'jsx-a11y/no-autofocus': 'error',
         'jsx-a11y/no-distracting-elements': 'error',
-        'jsx-a11y/no-noninteractive-element-handlers': 'error',
+        'jsx-a11y/no-noninteractive-element-interactions': 'error',
         'jsx-a11y/no-onchange': 'error',
         'jsx-a11y/no-redundant-roles': 'error',
         'jsx-a11y/no-static-element-interactions': 'warn',
